+++ src/Theatre.Application/Features/Halls/Commands/DeleteHall.cs
@@ -4,7 +4,7 @@ using Theatre.CqrsMediator.Special;
 
 namespace Theatre.Application.Features.Halls.Commands;
 
-public record DeleteHallCommand(short HallId) : IReturnType;
+public record DeleteHallCommand(int HallId) : IReturnType;
 
 public class DeleteHall(IHallsRepository hallsRepository) : ICommandHandler<DeleteHallCommand>
 {
